Refacto: 도메인 기반으로 파일 리팩토링
