+++ nanum_back/event-func/cobuying/dao/queryCoBuyingDetailDAO.ts
