+++ nanum_back/event-func/cobuying/dao/queryCoBuyingOwnerDAO.ts
