+++ nanum_back/event-func/cobuying/service/queryCoBuyingListSRV.ts
@@ -1,0 +1,27 @@
+import { queryCoBuyingListDAO } from '@cobuying/dao/queryCoBuyingListDAO';
+import { settingPageingQuery } from '@cobuying/service/settingPageingQuery';
+import { CoBuyingPageingRes, CoBuyingQueryParams, PageingQuery } from '@interface/cobuyingList';
+// import { CoBuyingSummaryProjectionExpression } from 'mappers/mapCoBuyingList';
+
+export const queryCoBuyingListSRV = async (input: CoBuyingQueryParams): Promise<CoBuyingPageingRes> => {
+    const query: PageingQuery = {
+        TableName: process.env.CoBuyingTableName || 'Dev-CoBuyingTable', // 테이블 이름
+        Limit: input.size || 20, // 최대 개수 (기본값 20)
+        IndexName: 'PagenationIndex',
+        KeyConditionExpression: 'deletedYN = :deletedYN',
+        ExpressionAttributeValues: {
+            ':deletedYN': { S: 'N' },
+        },
+        // CoBuyingSimple을 출력하게 기본 속성 정의. 정렬이나 필터링할 때 기준 속성 추가
+        ScanIndexForward: false,
+        // ProjectionExpression: CoBuyingSummaryProjectionExpression,
+    };
+
+    settingPageingQuery(input, query);
+    // 정렬기준에 따라 호출하는 쿼리가 달라짐
+    console.log('query : ', query);
+    // 생성일자 기준
+    const response: CoBuyingPageingRes = await queryCoBuyingListDAO(query);
+    return response;
+    // return response;
+};
