+++ nanum_back/event-func/common/optionsHandler.ts
@@ -1,0 +1,22 @@
+import { APIGatewayProxyEventV2, APIGatewayProxyResult } from "aws-lambda";
+import { LambdaReturnDto } from "@common/LambdaReturnDto";
+
+export const optionsHandler = async (event:APIGatewayProxyEventV2): Promise<APIGatewayProxyResult> => {
+  // return {
+  //   statusCode: 200,
+  //   headers: {
+  //     'Access-Control-Allow-Headers':
+  //         'Content-Type, Set-Cookie, x-amzn-Remapped-Authorization, Authorization, X-Forwarded-For, X-Api-Key, X-Amz-Security-Token, GongGong99-AccessToken, GongGong99-RefreshToken',
+  //       'Access-Control-Allow-Origin': event.headers.origin||'https://gonggong99.store' , // Allow from anywhere
+  //       'Access-Control-Allow-Methods': 'POST, GET, PUT, DELETE, OPTIONS', // Allow only GET request
+  //       'Access-Control-Allow-Credentials': 'true',
+  //       'Access-Control-Expose-Headers':
+  //         'Content-Type, x-amzn-Remapped-Authorization, Authorization, X-Forwarded-For, X-Api-Key, X-Amz-Security-Token, GongGong99-AccessToken, GongGong99-RefreshToken, Set-Cookie',
+  //   },
+  //   body: ""
+  // };
+
+    const dto = new LambdaReturnDto(200, {}, event).getLambdaReturnDto();
+    console.log("dto: "+ JSON.stringify(dto));
+    return dto;
+};
