+++ nanum_back/event-func/product/service/extractProductInfoSRV.ts
